#19 Rename module (deepl → translate)
