+++ tests/test_translate.py
@@ -1,5 +1,5 @@
 # third-party packages
-from astro_ph.deepl import Language, translate
+from astro_ph.translate import Language, translate
 from typing_extensions import Final
